requirements.txt
grpcio>=1.4.0,<1.5.0
python-consul
pytest
pytest-cov
flake8
coveralls
codeclimate-test-reporter
